Go back to lazy lookup, since at parse time it costs 5-8%
We still keep the super semantics in the message send node, instead of reading the argument. It fits better there, and is in sync with TruffleSOM.

This introduces a quickened super bytecode to use the inline cache.

I haven’t used a new node, since thought for the AST interpreter, which isn’t very consistent.

Signed-off-by: Stefan Marr <git@stefan-marr.de>