Introduce special calling convention for unary, binary, and ternary calls
This is hopefully reducing the amount of intermediate data structures needed during a call further, and specialized the code in the interpreter already a little.
Though, further interpreter work is needed to avoid the arity tests.

This simplified also block activation in while loops, and some other primitives where a new block was instatiated even though that's not needed, and even incorrect.

This also simplifies the signature creation of block methods, by using an array with the desired names.

Signed-off-by: Stefan Marr <git@stefan-marr.de>