Make message dispatch iterative
This possibly saves stack space.

There is also the simplification of the guard, and reducing number of times the rcvr_class needs to be determined.

The Generic Dispatch is now encoded also by the expected_class, as a simple guard after the dispatch loop. This avoids having to use a polymorphic guard checking method. Though, it's not necessarily obvious.

Signed-off-by: Stefan Marr <git@stefan-marr.de>