Remove stack from frame, and keep it mostly in interpreter()
This design is very similar to the bytecode interpreter in TruffleSOM.
We don't need to have the stack in the frame, it's just complexity there.

However, having it in the interpreter is a little ackward, too, because of all the code.
Though, it's trivial code, so, hopefully not much an issue.

For n-ary calls, we pass the stack, and stack pointer, and then callee can take what it needs.

Signed-off-by: Stefan Marr <git@stefan-marr.de>