Added bytecodes that push know constants, and specific constants
The bytecode profile indicates only 0 and 1 as common literals.
There are a few others, but those are really benchmark specific.
I added `nil` even though it's not super common, but it may be useful.

Give the varity of constants that are pushed, having push_constant_0..2 seems useful as a short cut and shorter version of the standard bytecode.

Signed-off-by: Stefan Marr <git@stefan-marr.de>