Add optimized LiteralReturn method
This is an optimization for trivial methods, to avoid the overhead of frame creation.
In TruffleSOM, this is done as part of the dispatch chain, which contains certain expressions.
Here, we do this as a different type of method, which should be easier applicable to both types of interpreters and avoids making the dispatch chain more complex.

Signed-off-by: Stefan Marr <git@stefan-marr.de>