Implement inlining of #while* loop messages in bytecode interpreter
Adapt heap_queue to use a class, because RPython doesn't support different tuple length, and field names are soo much more readable.

This also aborts optimizing dup-pop-pop sequences while inlining, they are already optimized, and we don't want to mess with branch targets.

First check number of arguments before doing a string compare in the parser to perhaps avoid unnecessary string compares.

This introduces the jump_backward bytecode.
During inlining, we can have bytecodes changing length, thus, we keep meta data on the loops in the methods, which is used to patch the branches up correctly.
We simply note when we see the start of a loop, not the bytecode index of the target method, and use that to determine the jump offset when emitting the jump_backward.

For the jit compiler, I moved the can_enter_jit to the backwards branch.
Need still to see whether this works well.

Signed-off-by: Stefan Marr <git@stefan-marr.de>