Add a minimal DSL for testing generated bytecode
The DSL is implemented in the check function.
A tuple is a bytecode with the address, all other bytecodes are assumed to come in linear order.

BC is a bytecode with arguments, and possibly a note.

BcMethod gets a get_bytecodes() method to make testing easier for embedded block methods.

Signed-off-by: Stefan Marr <git@stefan-marr.de>