Added inlining of #to:do: messages and handling for block arguments during inling
Block arguments are turned into locals during inlining.
This might not be perfect for the general case, but works well for to:do:.

This commit also adapts the variable write nodes so that I can pass in the value directly, and don't have to have a child node.
This requires some adaptation in the node traversal methods in abstract_node.py to handle the case that child nodes may be legitimately None.

This also moves the lexical scope into the abstract method.
Drawback is that also trivial methods have the extra field.

Signed-off-by: Stefan Marr <git@stefan-marr.de>