Revise ToDoInlined to prevent unnecessary store to frame
The biggest drawback of this design is that the lifetime of the frame is longer than the block activation in the loop, which is a problem for performance, because the compiler materializes the store.

Using `if we_are_jitted()` prevents the unnecessary store in the trace, and at the same time, avoids much of the overhead in the interpreter.

Signed-off-by: Stefan Marr <git@stefan-marr.de>